+++ packages/examples/packages/send-flow/src/components/SendFlow.tsx
@@ -15,9 +15,8 @@ import type { Account, Currency } from '../types';
  * @property selectedCurrency - The selected currency to display.
  * @property total - The total cost of the transaction.
  * @property fees - The fees for the transaction.
- * @property displayClearIcon - Whether to display the clear icon or not.
- * @property flushToAddress - Whether to flush the address field or not.
  * @property errors - The form errors.
+ * @property displayAvatar - Whether to display the avatar of the address.
  */
 export type SendFlowProps = {
   accounts: Account[];
@@ -25,12 +24,11 @@ export type SendFlowProps = {
   selectedCurrency: 'BTC' | '$';
   total: Currency;
   fees: Currency;
-  displayClearIcon: boolean;
-  flushToAddress?: boolean;
   errors?: {
     amount?: string;
     to?: string;
   };
+  displayAvatar?: boolean | undefined;
 };
 
 /**
@@ -43,8 +41,7 @@ export type SendFlowProps = {
  * @param props.total - The total cost of the transaction.
  * @param props.errors - The form errors.
  * @param props.fees - The fees for the transaction.
- * @param props.displayClearIcon - Whether to display the clear icon or not.
- * @param props.flushToAddress - Whether to flush the address field or not.
+ * @param props.displayAvatar - Whether to display the avatar of the address.
  * @returns The SendFlow component.
  */
 export const SendFlow: SnapComponent<SendFlowProps> = ({
@@ -53,9 +50,8 @@ export const SendFlow: SnapComponent<SendFlowProps> = ({
   selectedCurrency,
   total,
   fees,
-  displayClearIcon,
-  flushToAddress,
   errors,
+  displayAvatar,
 }) => {
   return (
     <Container>
@@ -65,9 +61,8 @@ export const SendFlow: SnapComponent<SendFlowProps> = ({
           selectedAccount={selectedAccount}
           accounts={accounts}
           selectedCurrency={selectedCurrency}
-          flushToAddress={flushToAddress}
-          displayClearIcon={displayClearIcon}
           errors={errors}
+          displayAvatar={displayAvatar}
         />
         <TransactionSummary fees={fees} total={total} />
       </Box>