+++ packages/snaps-rpc-methods/jest.config.js
@@ -10,10 +10,10 @@ module.exports = deepmerge(baseConfig, {
   ],
   coverageThreshold: {
     global: {
-      branches: 95.19,
-      functions: 98.62,
-      lines: 98.82,
-      statements: 98.5,
+      branches: 94.98,
+      functions: 98.64,
+      lines: 98.76,
+      statements: 98.45,
     },
   },
 });