+++ packages/snaps-rpc-methods/src/restricted/getEntropy.test.ts
@@ -1,8 +1,10 @@
 import { PermissionType, SubjectType } from '@metamask/permission-controller';
 import {
   MOCK_SNAP_ID,
-  TEST_SECRET_RECOVERY_PHRASE_BYTES,
+  TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
 } from '@metamask/snaps-utils/test-utils';
+import { hmac } from '@noble/hashes/hmac';
+import { sha512 } from '@noble/hashes/sha512';
 
 import { getEntropyBuilder } from './getEntropy';
 
@@ -12,7 +14,7 @@ describe('getEntropyBuilder', () => {
       targetName: 'snap_getEntropy',
       specificationBuilder: expect.any(Function),
       methodHooks: {
-        getMnemonic: true,
+        getMnemonicSeed: true,
         getUnlockPromise: true,
         getClientCryptography: true,
       },
@@ -21,7 +23,7 @@ describe('getEntropyBuilder', () => {
 
   it('returns the expected specification', () => {
     const methodHooks = {
-      getMnemonic: jest.fn(),
+      getMnemonicSeed: jest.fn(),
       getUnlockPromise: jest.fn(),
       getClientCryptography: jest.fn(),
     };
@@ -40,15 +42,15 @@ describe('getEntropyBuilder', () => {
 
 describe('getEntropyImplementation', () => {
   it('returns the expected result', async () => {
-    const getMnemonic = jest
+    const getMnemonicSeed = jest
       .fn()
-      .mockImplementation(() => TEST_SECRET_RECOVERY_PHRASE_BYTES);
+      .mockImplementation(() => TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES);
 
     const getUnlockPromise = jest.fn();
     const getClientCryptography = jest.fn().mockReturnValue({});
 
     const methodHooks = {
-      getMnemonic,
+      getMnemonicSeed,
       getUnlockPromise,
       getClientCryptography,
     };
@@ -74,15 +76,15 @@ describe('getEntropyImplementation', () => {
   });
 
   it('calls `getMnemonic` with a different entropy source', async () => {
-    const getMnemonic = jest
+    const getMnemonicSeed = jest
       .fn()
-      .mockImplementation(() => TEST_SECRET_RECOVERY_PHRASE_BYTES);
+      .mockImplementation(() => TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES);
 
     const getUnlockPromise = jest.fn();
     const getClientCryptography = jest.fn().mockReturnValue({});
 
     const methodHooks = {
-      getMnemonic,
+      getMnemonicSeed,
       getUnlockPromise,
       getClientCryptography,
     };
@@ -107,22 +109,26 @@ describe('getEntropyImplementation', () => {
       '0x6d8e92de419401c7da3cedd5f60ce5635b26059c2a4a8003877fec83653a4921',
     );
 
-    expect(getMnemonic).toHaveBeenCalledWith('source-id');
+    expect(getMnemonicSeed).toHaveBeenCalledWith('source-id');
   });
 
   it('uses custom client cryptography functions', async () => {
     const getUnlockPromise = jest.fn().mockResolvedValue(undefined);
-    const getMnemonic = jest
+    const getMnemonicSeed = jest
       .fn()
-      .mockResolvedValue(TEST_SECRET_RECOVERY_PHRASE_BYTES);
+      .mockResolvedValue(TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES);
 
-    const pbkdf2Sha512 = jest.fn().mockResolvedValue(new Uint8Array(64));
+    const hmacSha512 = jest
+      .fn()
+      .mockImplementation((key: Uint8Array, data: Uint8Array) =>
+        hmac(sha512, key, data),
+      );
     const getClientCryptography = jest.fn().mockReturnValue({
-      pbkdf2Sha512,
+      hmacSha512,
     });
 
     const methodHooks = {
-      getMnemonic,
+      getMnemonicSeed,
       getUnlockPromise,
       getClientCryptography,
     };
@@ -143,9 +149,9 @@ describe('getEntropyImplementation', () => {
     });
 
     expect(result).toBe(
-      '0x9bea47f2180fd874147f2f455a5ccc779826cfeff005605190cf0c568b3de7b5',
+      '0x6d8e92de419401c7da3cedd5f60ce5635b26059c2a4a8003877fec83653a4921',
     );
 
-    expect(pbkdf2Sha512).toHaveBeenCalledTimes(1);
+    expect(hmacSha512).toHaveBeenCalledTimes(10);
   });
 });