+++ packages/snaps-sdk/src/jsx/components/form/index.ts
@@ -1,3 +1,5 @@
+import type { AddressInputElement } from './AddressInput';
+import type { AssetSelectorElement } from './AssetSelector';
 import type { ButtonElement } from './Button';
 import type { CheckboxElement } from './Checkbox';
 import type { DropdownElement } from './Dropdown';
@@ -11,6 +13,7 @@ import type { RadioGroupElement } from './RadioGroup';
 import type { SelectorElement } from './Selector';
 import type { SelectorOptionElement } from './SelectorOption';
 
+export * from './AssetSelector';
 export * from './Button';
 export * from './Checkbox';
 export * from './Dropdown';
@@ -23,8 +26,11 @@ export * from './Form';
 export * from './Input';
 export * from './Selector';
 export * from './SelectorOption';
+export * from './AddressInput';
 
 export type StandardFormElement =
+  | AssetSelectorElement
+  | AddressInputElement
   | ButtonElement
   | CheckboxElement
   | FormElement