+++ packages/examples/packages/send-flow/CHANGELOG.md
@@ -7,11 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1]
+
+### Changed
+
+- Use new `displayAvatar` prop on `AddressInput` component ([#3264](https://github.com/MetaMask/snaps/pull/3264), [#3267](https://github.com/MetaMask/snaps/pull/3267), [#3268](https://github.com/MetaMask/snaps/pull/3268))
+
+## [1.1.0]
+
+### Changed
+
+- Use new `AddressInput` component ([#3167](https://github.com/MetaMask/snaps/pull/3167))
+
 ## [1.0.0]
 
 ### Changed
 
 - Initial release of the `@metamask/send-flow-example-snap` example.
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/send-flow-example-snap@1.0.0...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/send-flow-example-snap@1.1.1...HEAD
+[1.1.1]: https://github.com/MetaMask/snaps/compare/@metamask/send-flow-example-snap@1.1.0...@metamask/send-flow-example-snap@1.1.1
+[1.1.0]: https://github.com/MetaMask/snaps/compare/@metamask/send-flow-example-snap@1.0.0...@metamask/send-flow-example-snap@1.1.0
 [1.0.0]: https://github.com/MetaMask/snaps/releases/tag/@metamask/send-flow-example-snap@1.0.0